core_requirements.txt
## All required modules to install ##

# Use the following command to install these modules:
# pip3 install -r core_requirements.txt

python-telegram-bot
notify2
requests
tqdm
